Isolate other test_wms_getprint functions
