+++ tests/src/python/test_qgsserver_wms_getprint.py
@@ -24,7 +24,6 @@ import urllib.parse
 from qgis.testing import unittest
 
 from test_qgsserver import QgsServerTestBase
-from qgis.server import QgsServerRequest
 
 
 class TestQgsServerWMSGetPrint(QgsServerTestBase):
@@ -309,120 +308,6 @@ class TestQgsServerWMSGetPrint(QgsServerTestBase):
 
         r, h = self._result(self._execute_request(qs))
         self._img_diff_error(r, h, "WMS_GetPrint_Rotation")
-
-    def test_wms_getprint_selection(self):
-        qs = "?" + "&".join(["%s=%s" % i for i in list({
-            "MAP": urllib.parse.quote(self.projectPath),
-            "SERVICE": "WMS",
-            "VERSION": "1.1.1",
-            "REQUEST": "GetPrint",
-            "TEMPLATE": "layoutA4",
-            "FORMAT": "png",
-            "LAYERS": "Country,Hello",
-            "map0:EXTENT": "-33626185.498,-13032965.185,33978427.737,16020257.031",
-            "map0:LAYERS": "Country,Hello",
-            "CRS": "EPSG:3857",
-            "SELECTION": "Country: 4"
-        }.items())])
-
-        r, h = self._result(self._execute_request(qs))
-        self._img_diff_error(r, h, "WMS_GetPrint_Selection")
-
-    def test_wms_getprint_opacity(self):
-        qs = "?" + "&".join(["%s=%s" % i for i in list({
-            "MAP": urllib.parse.quote(self.projectPath),
-            "SERVICE": "WMS",
-            "VERSION": "1.1.1",
-            "REQUEST": "GetPrint",
-            "TEMPLATE": "layoutA4",
-            "FORMAT": "png",
-            "map0%3AEXTENT": "-33626185.498,-13032965.185,33978427.737,16020257.031",
-            "map0:LAYERS": "Country,Hello",
-            "CRS": "EPSG:3857",
-            "SELECTION": "Country: 4",
-            "LAYERS": "Country,Hello",
-            "OPACITIES": "125,125"
-        }.items())])
-
-        r, h = self._result(self._execute_request(qs))
-        self._img_diff_error(r, h, "WMS_GetPrint_Opacity")
-
-    def test_wms_getprint_opacity_post(self):
-        qs = "&".join(["%s=%s" % i for i in list({
-            "MAP": urllib.parse.quote(self.projectPath),
-            "SERVICE": "WMS",
-            "VERSION": "1.1.1",
-            "REQUEST": "GetPrint",
-            "TEMPLATE": "layoutA4",
-            "FORMAT": "png",
-            "map0%3AEXTENT": "-33626185.498,-13032965.185,33978427.737,16020257.031",
-            "map0:LAYERS": "Country,Hello",
-            "CRS": "EPSG:3857",
-            "SELECTION": "Country: 4",
-            "LAYERS": "Country,Hello",
-            "OPACITIES": "125%2C125"
-        }.items())])
-
-        r, h = self._result(self._execute_request('', QgsServerRequest.PostMethod, data=qs.encode('utf-8')))
-        self._img_diff_error(r, h, "WMS_GetPrint_Opacity")
-
-    def test_wms_getprint_highlight(self):
-        # default style
-        qs = "?" + "&".join(["%s=%s" % i for i in list({
-            "MAP": urllib.parse.quote(self.projectPath),
-            "SERVICE": "WMS",
-            "VERSION": "1.1.1",
-            "REQUEST": "GetPrint",
-            "TEMPLATE": "layoutA4",
-            "FORMAT": "png",
-            "map0:EXTENT": "-33626185.498,-13032965.185,33978427.737,16020257.031",
-            "map0:LAYERS": "Country_Labels",
-            "map0:HIGHLIGHT_GEOM": "POLYGON((-15000000 10000000, -15000000 6110620, 2500000 6110620, 2500000 10000000, -15000000 10000000))",
-            "map0:HIGHLIGHT_SYMBOL": "<StyledLayerDescriptor><UserStyle><Name>Highlight</Name><FeatureTypeStyle><Rule><Name>Symbol</Name><LineSymbolizer><Stroke><SvgParameter name=\"stroke\">%23ea1173</SvgParameter><SvgParameter name=\"stroke-opacity\">1</SvgParameter><SvgParameter name=\"stroke-width\">1.6</SvgParameter></Stroke></LineSymbolizer></Rule></FeatureTypeStyle></UserStyle></StyledLayerDescriptor>",
-            "map0:HIGHLIGHT_LABELSTRING": "Highlight Layer!",
-            "map0:HIGHLIGHT_LABELSIZE": "16",
-            "map0:HIGHLIGHT_LABELCOLOR": "%2300FF0000",
-            "map0:HIGHLIGHT_LABELBUFFERCOLOR": "%232300FF00",
-            "map0:HIGHLIGHT_LABELBUFFERSIZE": "1.5",
-            "CRS": "EPSG:3857"
-        }.items())])
-
-        r, h = self._result(self._execute_request(qs))
-        assert h.get("Content-Type").startswith('image'), r
-        self._img_diff_error(r, h, "WMS_GetPrint_Highlight")
-
-    def test_wms_getprint_label(self):
-        qs = "?" + "&".join(["%s=%s" % i for i in list({
-            "MAP": urllib.parse.quote(self.projectPath),
-            "SERVICE": "WMS",
-            "VERSION": "1.1.1",
-            "REQUEST": "GetPrint",
-            "TEMPLATE": "layoutA4",
-            "FORMAT": "png",
-            "map0:EXTENT": "-33626185.498,-13032965.185,33978427.737,16020257.031",
-            "map0:LAYERS": "Country,Hello",
-            "CRS": "EPSG:3857",
-            "IDTEXTBOX": "Updated QGIS composer label"
-        }.items())])
-
-        r, h = self._result(self._execute_request(qs))
-        self._img_diff_error(r, h, "WMS_GetPrint_LabelUpdated")
-
-        qs = "?" + "&".join(["%s=%s" % i for i in list({
-            "MAP": urllib.parse.quote(self.projectPath),
-            "SERVICE": "WMS",
-            "VERSION": "1.1.1",
-            "REQUEST": "GetPrint",
-            "TEMPLATE": "layoutA4",
-            "FORMAT": "png",
-            "map0:EXTENT": "-33626185.498,-13032965.185,33978427.737,16020257.031",
-            "map0:LAYERS": "Country,Hello",
-            "CRS": "EPSG:3857",
-            "IDTEXTBOX": ""
-        }.items())])
-
-        r, h = self._result(self._execute_request(qs))
-        self._img_diff_error(r, h, "WMS_GetPrint_LabelRemoved")
 
     def test_wms_getprint_two_maps(self):
         """Test map0 and map1 apply to the correct maps"""
